biomedhub/requirements.txt
djangorestframework
Django==1.4.5

